Update version to 0.1.1 and change dtype in training loop
This commit updates the project version in `pyproject.toml` to 0.1.1 and modifies the data type used in the `autocast` context of the training loop from `torch.float8_e4m3` to `torch.bfloat16`. These changes ensure compatibility with the latest training optimizations and improve numerical stability.

+++ src/splatnlp/model/training_loop.py
@@ -210,7 +210,7 @@ def train_epoch(
         optimizer.zero_grad()
 
         if scaler:
-            with autocast(device_type=config.device, dtype=torch.float8_e4m3):
+            with autocast(device_type=config.device, dtype=torch.bfloat16):
                 outputs = model(
                     batch_inputs,
                     batch_weapons,
